+++ app/Http/Controllers/cms/LocationController.php
@@ -137,7 +137,7 @@ class LocationController extends Controller
         
    public function locationtrashList()
    {
-        // $this->authorize('view', new Location());
+        $this->authorize('view', new Location());
         $data['trashedLocations'] = Location::onlyTrashed()->get();
 
         return view('cms.trashed.locationTrashedList',$data);
@@ -146,7 +146,7 @@ class LocationController extends Controller
 
    public function restoreLocation($id)
    {
-        // $this->authorize('restore', new Location());
+        $this->authorize('restore', new Location());
         $location = Location::onlyTrashed()->find($id)->restore();
  
         return back()->with('success','Successfully Restored');
@@ -155,7 +155,7 @@ class LocationController extends Controller
 
    public function forceDeleteLocation($id)
    {
-        // $this->authorize('forceDelete',new Location);
+        $this->authorize('forceDelete',new Location);
         $location = Location::onlyTrashed()->find($id)->forceDelete();
  
         return back()->with('success','Permanently Deleted');
